Announce the competition
This is based on a combination of the Tech Day 2 email and last
year's competition/tech day 3 email.

This is very much a first draft, and still has some missing pieces.

Patch set #2: emdash and s/yourselves/you/ in one case
Patch set #3: remove emdash so that the email sending system can
              actually manage the content of the email

Change-Id: I92541f758909ec03ae7a28af87621b76925477ee